Revert "Function bugfix"

Accidentally merged into main, my bad.

+++ blueprint/lib/siteQueries.js
@@ -1,5 +1,5 @@
-import { commit, validateConnection } from "./utility.js";
-import { connection } from "./connection.js"
+import { closeConnection } from "./connection";
+import { commit } from "./utility";
 
 /**
     getSites() - Retrieves all sites from the database.
@@ -9,14 +9,9 @@ import { connection } from "./connection.js"
     Author: Lydell Jones
     Dependencies: mysql
 */
-export async function getSites() {
-    var result;
-    try {
-        await validateConnection();
-        [result] = await connection.query(`SELECT * FROM userWebsites;`);
-    } catch (err) {
-        throw undefined;
-    }
+export async function getSites(connection) {
+    if (!validateConnection(connection)) return false;
+    const [result] = await connection.query(`SELECT * FROM userWebsites;`);
     return result;
 }
 
@@ -28,17 +23,9 @@ export async function getSites() {
     Author: Lydell Jones
     Dependencies: mysql
 */
-export async function getSiteByID(id) {
-    var result;
-    try {
-        await validateConnection();
-        [result] = await connection.query(`SELECT * FROM userWebsites WHERE id = ?;`, [id]);
-    } catch (err) {
-        throw undefined;
-    }
-    return result;
-    // if (!await validateConnection()) return false;
-    // return await connection.query(`SELECT * FROM userWebsites WHERE id = ?;`, [id])
+export async function getSiteByID(id, connection) {
+    if (!await validateConnection(connection)) return false;
+    await connection.query(`SELECT * FROM userWebsites WHERE id = ?;`, [id])
 }
 
 /**
@@ -49,18 +36,10 @@ export async function getSiteByID(id) {
     Author: Lydell Jones
     Dependencies: mysql
 */
-export async function getSiteByName(name) {
-    var result;
-    try {
-        await validateConnection();
-        [result] = await connection.query(`SELECT * FROM userWebsites WHERE websiteName = ?;`, [name]);
-    } catch (err) {
-        throw undefined;
-    }
+export async function getSiteByName(name, connection) {
+    if (!await validateConnection(connection)) return false;
+    const [result] = await connection.query(`SELECT * FROM userWebsites WHERE websiteName = ?;`, [name]);
     return result;
-    // if (!await validateConnection()) return false;
-    // const [result] = await connection.query(`SELECT * FROM userWebsites WHERE websiteName = ?;`, [name]);
-    // return result;
 }
 
 /**
@@ -71,18 +50,10 @@ export async function getSiteByName(name) {
  * Author: Lydell Jones
  * Dependencies: mysql
  */
-export async function getSiteCount() {
-    var result;
-    try {
-        await validateConnection();
-        [result] = await connection.query(`SELECT COUNT(*) FROM userWebsites;`);
-    } catch (err) {
-        throw undefined;
-    }
-    return result;
-    // if (!await validateConnection()) return false;
-    // const [result] = await connection.query(`SELECT COUNT(*) FROM userWebsites;`)
-    // return result;
+export async function getSiteCount(connection) {
+    if (!await validateConnection(connection)) return false;
+    const [result] = await connection.query(`SELECT COUNT(*) FROM userWebsites;`)
+    return false;
 }
 
 /**
@@ -93,18 +64,10 @@ export async function getSiteCount() {
  * Author: Lydell Jones
  * Dependencies: mysql
  */
-export async function createSite(name, userId) {
-    try {
-        await validateConnection()
-        await connection.query(`INSERT INTO userWebsites (websiteName, website_user_id) VALUES (?, ?);`, [name, userId]);
-        await commit(connection);
-    } catch (err) {
-        throw false;
-    }
-    return true;
-    // if (!await validateConnection()) return false;
-    // await connection.query(`INSERT INTO userWebsites (websiteName, website_user_id) VALUES (?, ?);`, [name, userId]);
-    // return await commit(connection);
+export async function createSite(name, connection, userId) {
+    if (!await validateConnection(connection)) return false;
+    await connection.query(`INSERT INTO userWebsites (websiteName, website_user_id) VALUES (?, ?);`, [name, userId]);
+    return await commit(connection);
 }
 
 /**
@@ -115,18 +78,10 @@ export async function createSite(name, userId) {
  * Author: Lydell Jones
  * Dependencies: mysql
  */
-export async function deleteSite(siteId) {
-    try { 
-        await validateConnection();
-        await connection.query(`DELETE FROM userWebsites WHERE id = ?;`, [siteId]);
-        await commit();
-    } catch (err) {
-        throw false;
-    }
-    return true;
-    // if (!validateConnection()) return false;
-    // await connection.query(`DELETE FROM userWebsites WHERE id = ?;`, [siteId]);
-    // return await commit();
+export async function deleteSite(siteId, connection) {
+    if (!validateConnection(connection)) return false;
+    await connection.query(`DELETE FROM userWebsites WHERE id = ?;`, [siteId]);
+    return await commit(connection);
 }
 
 /**
@@ -138,17 +93,8 @@ export async function deleteSite(siteId) {
  * Author: Lydell Jones
  * Dependencies: mysql
  */
-export async function updateSite(id, name) {
-    try {
-        await validateConnection();
-        await connection.query(`UPDATE userWebsites SET websiteName=?, websiteDateLastModified=NOW()  WHERE id = ?;`, [name, id]);
-        await commit();
-    } catch (err) {
-        console.error(err);
-        throw false;
-    }
-    return true;
-    // if (!validateConnection()) return false;
-    // await connection.query(`UPDATE userWebsites SET websiteName=?, websiteDateLastModified=NOW()  WHERE id = ?;`, [name, id]);
-    // return await commit();
+export async function updateSite(id, name, connection) {
+    if (!validateConnection(connection)) return false;
+    await connection.query(`UPDATE userWebsites SET websiteName=?, websiteDateLastModified=NOW()  WHERE id = ?;`, [data, id]);
+    return await commit(connection);
 }
